refactor(components): move options/hooks to src/hooks/options
