+++ frontend/src/forms/MultiSelectFormField.tsx
@@ -20,7 +20,7 @@ import { Control, Controller, FieldValues, Path } from "react-hook-form";
 import { MultiValue } from "react-select";
 
 import MultiSelect from "@/components/ui/multi-select/MultiSelect";
-import type { Option } from "@/components/options/hooks";
+import type { Option } from "@/hooks/options";
 
 type MultiSelectValue = {
   id: string;
