+++ frontend/src/forms/SelectFormFIeld.tsx
@@ -19,7 +19,7 @@
 import { Control, Controller, FieldValues, Path } from "react-hook-form";
 import type { Props } from "react-select";
 
-import type { Option } from "@/components/options/hooks";
+import type { Option } from "@/hooks/options";
 import Select from "@/components/ui/select/Select";
 import { useMemo } from "react";
